+++ packages/overlays/src/configurations/withModalDialogConfig.js
@@ -1,0 +1,12 @@
+export const withModalDialogConfig = () => ({
+  placementMode: 'global',
+  viewportConfig: {
+    placement: 'center',
+  },
+
+  hasBackdrop: true,
+  preventsScroll: true,
+  trapsKeyboardFocus: true,
+  hidesOnEsc: true,
+  handlesAccessibility: true,
+});