+++ packages/tooltip/src/LionTooltip.js
@@ -9,7 +9,7 @@ export class LionTooltip extends LionPopup {
 
   connectedCallback() {
     super.connectedCallback();
-    this.contentNode.setAttribute('role', 'tooltip');
+    this._overlayContentNode.setAttribute('role', 'tooltip');
 
     this.__resetActive = () => {
       this.mouseActive = false;
@@ -19,42 +19,42 @@ export class LionTooltip extends LionPopup {
     this.__showMouse = () => {
       if (!this.keyActive) {
         this.mouseActive = true;
-        this._controller.show();
+        this._overlayCtrl.show();
       }
     };
 
     this.__hideMouse = () => {
       if (!this.keyActive) {
-        this._controller.hide();
+        this._overlayCtrl.hide();
       }
     };
 
     this.__showKey = () => {
       if (!this.mouseActive) {
         this.keyActive = true;
-        this._controller.show();
+        this._overlayCtrl.show();
       }
     };
 
     this.__hideKey = () => {
       if (!this.mouseActive) {
-        this._controller.hide();
+        this._overlayCtrl.hide();
       }
     };
 
-    this._controller.addEventListener('hide', this.__resetActive);
+    this._overlayCtrl.addEventListener('hide', this.__resetActive);
     this.addEventListener('mouseenter', this.__showMouse);
     this.addEventListener('mouseleave', this.__hideMouse);
-    this.invokerNode.addEventListener('focusin', this.__showKey);
-    this.invokerNode.addEventListener('focusout', this.__hideKey);
+    this._overlayInvokerNode.addEventListener('focusin', this.__showKey);
+    this._overlayInvokerNode.addEventListener('focusout', this.__hideKey);
   }
 
   disconnectedCallback() {
     super.disconnectedCallback();
-    this._controller.removeEventListener('hide', this.__resetActive);
+    this._overlayCtrl.removeEventListener('hide', this.__resetActive);
     this.removeEventListener('mouseenter', this.__showMouse);
     this.removeEventListener('mouseleave', this._hideMouse);
-    this.invokerNode.removeEventListener('focusin', this._showKey);
-    this.invokerNode.removeEventListener('focusout', this._hideKey);
+    this._overlayInvokerNode.removeEventListener('focusin', this._showKey);
+    this._overlayInvokerNode.removeEventListener('focusout', this._hideKey);
   }
 }