+++ packages/calendar/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.2.0](https://github.com/ing-bank/lion/compare/@lion/calendar@0.1.71...@lion/calendar@0.2.0) (2019-10-10)
+
+
+### Features
+
+* update to latest overlay system ([4c26bef](https://github.com/ing-bank/lion/commit/4c26bef))
+
+
+
+
+
+## [0.1.71](https://github.com/ing-bank/lion/compare/@lion/calendar@0.1.70...@lion/calendar@0.1.71) (2019-10-09)
+
+**Note:** Version bump only for package @lion/calendar
+
+
+
+
+
+## [0.1.70](https://github.com/ing-bank/lion/compare/@lion/calendar@0.1.69...@lion/calendar@0.1.70) (2019-10-07)
+
+**Note:** Version bump only for package @lion/calendar
+
+
+
+
+
 ## [0.1.69](https://github.com/ing-bank/lion/compare/@lion/calendar@0.1.68...@lion/calendar@0.1.69) (2019-09-30)
 
 **Note:** Version bump only for package @lion/calendar