+++ packages/radio/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.58](https://github.com/ing-bank/lion/compare/@lion/radio@0.1.57...@lion/radio@0.1.58) (2019-10-09)
+
+**Note:** Version bump only for package @lion/radio
+
+
+
+
+
+## [0.1.57](https://github.com/ing-bank/lion/compare/@lion/radio@0.1.56...@lion/radio@0.1.57) (2019-10-07)
+
+**Note:** Version bump only for package @lion/radio
+
+
+
+
+
 ## [0.1.56](https://github.com/ing-bank/lion/compare/@lion/radio@0.1.55...@lion/radio@0.1.56) (2019-09-30)
 
 **Note:** Version bump only for package @lion/radio